+++ src/main/java/cloud/prefab/client/ConfigStore.java
@@ -1,17 +1,12 @@
 package cloud.prefab.client;
 
-import cloud.prefab.domain.Prefab;
+import cloud.prefab.client.config.ConfigElement;
 import java.util.Collection;
-import java.util.Map;
-import java.util.Optional;
 
 public interface ConfigStore {
-  Optional<Prefab.ConfigValue> get(String key);
-
-  Optional<Prefab.ConfigValue> get(
-    String key,
-    Map<String, Prefab.ConfigValue> properties
-  );
-
   Collection<String> getKeys();
+
+  ConfigElement getElement(String key);
+
+  boolean containsKey(String key);
 }
